Moved TestListenerWithoutSite into listener_test.go file.
